requirements.txt
PyQt6>=6.5.0
opencv-python>=4.8.0.74
numpy>=1.24.0
requests>=2.31.0
Pillow>=10.0.0